+++ Util/SetupUtils/InstallPrerequisites.sh
@@ -36,6 +36,21 @@ if [ "$EUID" -ne 0 ]; then
     exit 1
 fi
 
+# -- DETECT UBUNTU VERSION --
+if [ -f /etc/os-release ]; then
+    . /etc/os-release
+    UBUNTU_VERSION_ID="${VERSION_ID}"
+else
+    UBUNTU_VERSION_ID="22.04"
+fi
+
+# Select libtiff package name based on Ubuntu version
+if dpkg --compare-versions "$UBUNTU_VERSION_ID" ge "24.04"; then
+    LIBTIFF_PKG="libtiff-dev"
+else
+    LIBTIFF_PKG="libtiff5-dev"
+fi
+
 # -- INSTALL APT PACKAGES --
 echo "Installing Ubuntu Packages..."
 apt-get update
@@ -45,7 +60,7 @@ apt-get -y install \
     ninja-build \
     libvulkan1 \
     libpng-dev \
-    libtiff5-dev \
+    "$LIBTIFF_PKG" \
     libjpeg-dev \
     tzdata \
     sed \
@@ -71,8 +86,12 @@ fi
 
 # -- INSTALL PYTHON PACKAGES --
 echo "Installing Python Packages..."
-$python_path -m pip install --upgrade pip
-$python_path -m pip install -r requirements.txt
+PIP_EXTRA_ARGS=""
+if dpkg --compare-versions "$UBUNTU_VERSION_ID" ge "24.04"; then
+    PIP_EXTRA_ARGS="--break-system-packages"
+fi
+$python_path -m pip install --upgrade pip $PIP_EXTRA_ARGS
+$python_path -m pip install -r requirements.txt $PIP_EXTRA_ARGS
 
 # -- INSTALL CMAKE --
 check_cmake_version() {